backend/requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Werkzeug==3.0.1
python-dotenv==1.0.0
openai==1.12.0
gunicorn==21.2.0
# psycopg2-binary==2.9.9
mysql-connector-python==8.3.0
flask-cors==4.0.0
pyjwt==2.8.0
